// finalAssesment/ProjectManagerWebApp/src/app/shared/models/tasks.model.ts
export class Tasks {
    Task_ID: number;
    Task: string;
    Project_ID: Number;
    Parent_ID: Number;
    Start_Date: Date;
    End_Date: Date;
    Priority: string;
    Name:string;
    User_ID:number;
    Project:string;
    ParentTask:string;
    IsParentTask:boolean;
    TaskStatus:string;
}